+++ tools/visual-pipeline-and-platform-evaluation-tool/ui/src/pages/Videos.tsx
@@ -15,53 +15,18 @@ import {
 } from "@/components/ui/table.tsx";
 import { formatElapsedTimeSeconds } from "@/lib/timeUtils.ts";
 import { filterOutTransportStreams } from "@/lib/videoUtils.ts";
-import { useForm } from "react-hook-form";
-import { Button } from "@/components/ui/button.tsx";
-import { Input } from "@/components/ui/input.tsx";
-import { Label } from "@/components/ui/label.tsx";
-import { useRef, useState, useEffect } from "react";
-import { AlertCircle, CheckCircle2, Loader2, Upload, X } from "lucide-react";
-import { Progress } from "@/components/ui/progress.tsx";
+import { useEffect, useCallback } from "react";
 import { useAppDispatch } from "@/store/hooks";
 import { toast } from "sonner";
 import { useBackgroundJobs } from "@/contexts/useBackgroundJobs";
-
-type UploadFormData = {
-  files: FileList | null;
-};
-
-type FileUploadState = {
-  file: File;
-  status: "pending" | "uploading" | "completed" | "failed";
-  progress: number;
-  error?: string;
-};
-
-type FileUploadJob = FileUploadState & {
-  originalIndex: number;
-};
-
-const MAX_CONCURRENT_UPLOADS = 3;
+import { MultiFileUploader } from "@/components/shared/MultiFileUploader.tsx";
 
 export const Videos = () => {
-  const { data: videos, isSuccess } = useGetVideosQuery();
+  const { data: videos, isSuccess, isLoading } = useGetVideosQuery();
   const dispatch = useAppDispatch();
   const [checkVideoExists] = useLazyCheckVideoInputExistsQuery();
   const { registerJobGroup, unregisterJobGroup, updateJobs } =
     useBackgroundJobs();
-  const { register, handleSubmit, reset, watch, setValue } =
-    useForm<UploadFormData>({
-      defaultValues: {
-        files: null,
-      },
-    });
-
-  const [isDragging, setIsDragging] = useState(false);
-  const [selectedFilesList, setSelectedFilesList] = useState<File[]>([]);
-  const [uploadStates, setUploadStates] = useState<FileUploadState[]>([]);
-  const [isUploading, setIsUploading] = useState(false);
-  const [isPostUpload, setIsPostUpload] = useState(false);
-  const fileInputRef = useRef<HTMLInputElement | null>(null);
 
   // Register this component as a job group
   useEffect(() => {
@@ -71,653 +36,75 @@ export const Videos = () => {
     };
   }, [registerJobGroup, unregisterJobGroup]);
 
-  // Update global upload progress whenever uploadStates changes
-  useEffect(() => {
-    const jobs = uploadStates
-      .filter((state) => state.error !== "File already exists on server")
-      .map((state, index) => ({
-        id: `video-${index}-${state.file.name}`,
-        name: state.file.name,
-        progress: state.progress,
-      }));
-
-    updateJobs("videos", jobs);
-  }, [uploadStates, updateJobs]);
-
-  const selectedFiles = watch("files");
-  const fileCount = selectedFiles?.length || selectedFilesList.length;
-
-  // Calculate overall progress
-  // Calculate overall progress only for uploadable files (exclude files that already exist)
-  const uploadableStates = uploadStates.filter(
-    (s) => s.error !== "File already exists on server",
+  const handleCheckFileExists = useCallback(
+    async (filename: string): Promise<{ exists: boolean }> => {
+      try {
+        const result = await checkVideoExists({ filename }).unwrap();
+        return { exists: result.exists };
+      } catch (error) {
+        console.error(`Error checking file ${filename}:`, error);
+        return { exists: false };
+      }
+    },
+    [checkVideoExists],
   );
-  const overallProgress =
-    uploadableStates.length > 0
-      ? uploadableStates.reduce((sum, state) => sum + state.progress, 0) /
-        uploadableStates.length
-      : 0;
 
-  // Count only files that will be uploaded (exclude files that already exist)
-  const uploadableFilesCount = uploadStates.filter(
-    (s) => s.error !== "File already exists on server",
-  ).length;
-  const uploadableCompleted = uploadStates.filter(
-    (s) => s.status === "completed",
-  ).length;
-  const uploadableFailed = uploadStates.filter(
-    (s) => s.status === "failed" && s.error !== "File already exists on server",
-  ).length;
+  const handleUploadProgress = useCallback(
+    (jobs: Array<{ id: string; name: string; progress: number }>) => {
+      updateJobs("videos", jobs);
+    },
+    [updateJobs],
+  );
 
-  /**
-   * Check if files already exist on the backend
-   * Returns array of files with their existence status
-   */
-  const checkFilesExistence = async (
-    files: File[],
-  ): Promise<Array<{ file: File; exists: boolean }>> =>
-    await Promise.all(
-      files.map(async (file) => {
-        try {
-          const result = await checkVideoExists({
-            filename: file.name,
-          }).unwrap();
-          return { file, exists: result.exists };
-        } catch (error) {
-          // If check fails, assume file doesn't exist
-          console.error(`Error checking file ${file.name}:`, error);
-          return { file, exists: false };
-        }
-      }),
-    );
-
-  /**
-   * Upload a single file with progress tracking
-   *
-   * API IMPLEMENTATION HINTS:
-   *
-   * Backend endpoint: POST /api/v1/videos (configured in apiSlice baseUrl + endpoint)
-   *
-   * The endpoint should:
-   * 1. Accept multipart/form-data with field name "file"
-   * 2. Support chunked transfer encoding for progress tracking
-   * 3. Return appropriate HTTP status codes (201 for success, 4xx/5xx for errors)
-   * 4. Return Video object with metadata after processing
-   *
-   * Example backend (Express.js + Multer):
-   * ```javascript
-   * const multer = require('multer');
-   * const upload = multer({ dest: 'uploads/videos/' });
-   *
-   * app.post('/api/v1/videos', upload.single('file'), (req, res) => {
-   *   // req.file contains the uploaded file
-   *   // Validate file type, size, etc.
-   *   // Process the video (extract metadata, generate thumbnails, etc.)
-   *   res.status(201).json({
-   *     filename: req.file.originalname,
-   *     width: 1920,
-   *     height: 1080,
-   *     fps: 30,
-   *     frame_count: 300,
-   *     codec: 'h264',
-   *     duration: 10.0
-   *   });
-   * });
-   * ```
-   *
-   * Example backend (FastAPI):
-   * ```python
-   * from fastapi import FastAPI, File, UploadFile
-   *
-   * @app.post("/api/v1/videos")
-   * async def upload_video(file: UploadFile = File(...)):
-   *     # Save file
-   *     file_path = f"uploads/videos/{file.filename}"
-   *     with open(file_path, "wb") as f:
-   *         content = await file.read()
-   *         f.write(content)
-   *     # Extract video metadata using ffprobe or similar
-   *     return {
-   *         "filename": file.filename,
-   *         "width": 1920,
-   *         "height": 1080,
-   *         "fps": 30.0,
-   *         "frame_count": 300,
-   *         "codec": "h264",
-   *         "duration": 10.0
-   *     }
-   * ```
-   */
-  const uploadFile = async (
-    file: File,
-    onProgress: (progress: number) => void,
-  ): Promise<void> => {
-    return new Promise((resolve, reject) => {
-      const formData = new FormData();
-      formData.append("file", file); // Backend expects "file" field
-
-      // Use XMLHttpRequest for progress tracking
-      // Note: fetch() doesn't support upload progress natively
-      const xhr = new XMLHttpRequest();
-
-      // Track upload progress
-      xhr.upload.addEventListener("progress", (e) => {
-        if (e.lengthComputable) {
-          const progress = (e.loaded / e.total) * 100;
-          onProgress(progress);
-        }
-      });
-
-      // Handle completion
-      xhr.addEventListener("load", () => {
-        if (xhr.status >= 200 && xhr.status < 300) {
-          resolve();
-        } else {
-          try {
-            const response = JSON.parse(xhr.responseText);
-            const detail =
-              response.detail || `Upload failed with status ${xhr.status}`;
-            reject(new Error(detail));
-          } catch {
-            reject(new Error(`Upload failed with status ${xhr.status}`));
-          }
-        }
-      });
-
-      xhr.addEventListener("error", () => {
-        reject(new Error("Network error during upload"));
-      });
-
-      xhr.addEventListener("abort", () => {
-        reject(new Error("Upload aborted"));
-      });
-
-      // Configure and send request to the correct API endpoint
-      xhr.open("POST", ENDPOINTS.UPLOAD_VIDEO);
-      xhr.send(formData);
-    });
-  };
-
-  /**
-   * Process uploads with concurrency limit
-   * Uses a queue system to ensure max 3 concurrent uploads
-   */
-  const processUploadsWithConcurrency = async (
-    files: FileUploadJob[],
-  ): Promise<{ succeeded: number; failed: number }> => {
-    const executing = new Set<Promise<void>>();
-    let succeeded = 0;
-    let failed = 0;
-
-    for (const fileJob of files) {
-      const uploadPromise = async () => {
-        const { originalIndex } = fileJob;
-        try {
-          // Update status to uploading
-          setUploadStates((prev) => {
-            const newStates = [...prev];
-            if (newStates[originalIndex]) {
-              newStates[originalIndex] = {
-                ...newStates[originalIndex],
-                status: "uploading",
-              };
-            }
-            return newStates;
-          });
-
-          // Upload with progress tracking
-          await uploadFile(fileJob.file, (progress) => {
-            setUploadStates((prev) => {
-              const newStates = [...prev];
-              if (newStates[originalIndex]) {
-                newStates[originalIndex] = {
-                  ...newStates[originalIndex],
-                  progress,
-                };
-              }
-              return newStates;
-            });
-          });
-
-          // Mark as completed
-          setUploadStates((prev) => {
-            const newStates = [...prev];
-            if (newStates[originalIndex]) {
-              newStates[originalIndex] = {
-                ...newStates[originalIndex],
-                status: "completed",
-                progress: 100,
-              };
-            }
-            return newStates;
-          });
-          succeeded++;
-        } catch (error) {
-          // Mark as failed
-          setUploadStates((prev) => {
-            const newStates = [...prev];
-            if (newStates[originalIndex]) {
-              newStates[originalIndex] = {
-                ...newStates[originalIndex],
-                status: "failed",
-                error: error instanceof Error ? error.message : "Upload failed",
-              };
-            }
-            return newStates;
-          });
-          failed++;
-        }
-      };
-
-      const promise = uploadPromise().finally(() => {
-        executing.delete(promise);
-      });
-
-      executing.add(promise);
-
-      // Limit concurrency - wait if we have MAX_CONCURRENT_UPLOADS running
-      if (executing.size >= MAX_CONCURRENT_UPLOADS) {
-        await Promise.race(executing);
-      }
-    }
-
-    // Wait for remaining uploads
-    await Promise.all(executing);
-
-    return { succeeded, failed };
-  };
-
-  const handleDragEnter = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-
-  const handleDragLeave = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  };
-
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-  };
-
-  const handleDrop = async (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-
-    const droppedFiles = Array.from(e.dataTransfer.files).filter((file) =>
-      file.type.startsWith("video/"),
-    );
-
-    if (droppedFiles.length > 0) {
-      // Capture the post-upload state before resetting it
-      const shouldReplaceList = isPostUpload;
-
-      // If post-upload, replace the list; otherwise, add to existing list
-      const baseFiles = shouldReplaceList ? [] : selectedFilesList;
-
-      // Filter out duplicates by filename (files already in the list)
-      const existingFileNames = new Set(baseFiles.map((f) => f.name));
-      const newFiles = droppedFiles.filter(
-        (file) => !existingFileNames.has(file.name),
-      );
-
-      if (newFiles.length === 0) return;
-
-      // Check which files already exist on backend
-      const fileChecks = await checkFilesExistence(newFiles);
-
-      // Add all new files to the list (or replace if post-upload)
-      const allFiles = [...baseFiles, ...newFiles];
-      setSelectedFilesList(allFiles);
-
-      // Reset post-upload flag after first file selection
-      if (shouldReplaceList) {
-        setIsPostUpload(false);
-      }
-
-      // Create a new FileList-like object for react-hook-form
-      const dataTransfer = new DataTransfer();
-      allFiles.forEach((file) => dataTransfer.items.add(file));
-      setValue("files", dataTransfer.files);
-
-      // Initialize upload states for new files (or replace if post-upload)
-      const newUploadStates: FileUploadState[] = fileChecks.map(
-        ({ file, exists }) => ({
-          file,
-          status: exists ? "failed" : "pending",
-          progress: 0,
-          error: exists ? "File already exists on server" : undefined,
-        }),
-      );
-
-      setUploadStates(
-        shouldReplaceList
-          ? newUploadStates
-          : [...uploadStates, ...newUploadStates],
-      );
-    }
-  };
-
-  const handleFileInputChange = async (
-    e: React.ChangeEvent<HTMLInputElement>,
-  ) => {
-    const files = e.target.files;
-    if (files && files.length > 0) {
-      // Capture the post-upload state before resetting it
-      const shouldReplaceList = isPostUpload;
-
-      // If post-upload, replace the list; otherwise, add to existing list
-      const baseFiles = shouldReplaceList ? [] : selectedFilesList;
-
-      // Filter out duplicates by filename (files already in the list)
-      const existingFileNames = new Set(baseFiles.map((f) => f.name));
-      const newFiles = Array.from(files).filter(
-        (file) => !existingFileNames.has(file.name),
-      );
-
-      if (newFiles.length === 0) return;
-
-      // Check which files already exist on backend
-      const fileChecks = await checkFilesExistence(newFiles);
-
-      // Add all new files to the list (or replace if post-upload)
-      const allFiles = [...baseFiles, ...newFiles];
-      setSelectedFilesList(allFiles);
-
-      // Reset post-upload flag after first file selection
-      if (shouldReplaceList) {
-        setIsPostUpload(false);
-      }
-
-      // Create a new FileList-like object for react-hook-form
-      const dataTransfer = new DataTransfer();
-      allFiles.forEach((file) => dataTransfer.items.add(file));
-      setValue("files", dataTransfer.files);
-
-      // Initialize upload states for new files (or replace if post-upload)
-      const newUploadStates: FileUploadState[] = fileChecks.map(
-        ({ file, exists }) => ({
-          file,
-          status: exists ? "failed" : "pending",
-          progress: 0,
-          error: exists ? "File already exists on server" : undefined,
-        }),
-      );
-
-      setUploadStates(
-        shouldReplaceList
-          ? newUploadStates
-          : [...uploadStates, ...newUploadStates],
-      );
-    }
-  };
-
-  const removeFile = (index: number) => {
-    const newFiles = selectedFilesList.filter((_, i) => i !== index);
-    const newUploadStates = uploadStates.filter((_, i) => i !== index);
-    setSelectedFilesList(newFiles);
-    setUploadStates(newUploadStates);
-
-    if (newFiles.length === 0) {
-      setValue("files", null);
-      if (fileInputRef.current) {
-        fileInputRef.current.value = "";
-      }
-    } else {
-      const dataTransfer = new DataTransfer();
-      newFiles.forEach((file) => dataTransfer.items.add(file));
-      setValue("files", dataTransfer.files);
-    }
-  };
-
-  const onSubmit = async () => {
-    if (selectedFilesList.length === 0 || isUploading) {
-      return;
-    }
-
-    setIsUploading(true);
-
-    // Filter out files that already failed (exist on server)
-    // and only initialize upload states for files that need to be uploaded
-    // Keep track of original indices to update the correct items in uploadStates
-    const filesToUpload: FileUploadJob[] = uploadStates
-      .map((state, originalIndex) => ({ ...state, originalIndex }))
-      .filter((state) => state.status === "pending");
-
-    // If there are no files to upload (all failed checks), just stop
-    if (filesToUpload.length === 0) {
-      setIsUploading(false);
-      return;
-    }
-
-    try {
-      // Process uploads with concurrency control
-      const { succeeded, failed } =
-        await processUploadsWithConcurrency(filesToUpload);
-
-      // Check if all uploadable files succeeded
-      if (failed === 0 && succeeded === filesToUpload.length) {
-        // Invalidate videos cache to refetch the updated list
+  const handleUploadComplete = useCallback(
+    (succeeded: number, failed: number) => {
+      if (failed === 0 && succeeded > 0) {
         dispatch(api.util.invalidateTags(["videos"]));
-
-        // Show success toast
         toast.success("Upload completed.");
-
-        // Reset form after successful upload
-        setTimeout(() => {
-          setSelectedFilesList([]);
-          setUploadStates([]);
-          reset();
-          if (fileInputRef.current) {
-            fileInputRef.current.value = "";
-          }
-          setIsUploading(false);
-          setIsPostUpload(true);
-        }, 2000); // Show success state for 2 seconds
-      } else {
-        setIsUploading(false);
-        setIsPostUpload(true);
+      } else if (succeeded > 0 && failed > 0) {
+        toast.warning(
+          `${succeeded} file(s) uploaded successfully. ${failed} failed.`,
+        );
+        dispatch(api.util.invalidateTags(["videos"]));
+      } else if (failed > 0) {
+        toast.error(`Upload failed for ${failed} file(s).`);
       }
-    } catch (error) {
-      console.error("Upload error:", error);
-      setIsUploading(false);
-      setIsPostUpload(true);
-    }
-  };
+    },
+    [dispatch],
+  );
+
   const filteredVideos =
     isSuccess && videos ? filterOutTransportStreams(videos) : [];
 
-  if (isSuccess && filteredVideos.length > 0) {
+  if (isLoading) {
     return (
-      <div className="container pl-16 mx-auto py-10">
-        <div className="mb-6">
-          <h1 className="text-3xl font-bold">Videos</h1>
-          <p className="text-muted-foreground mt-2">
-            Ready-to-use video clips available in the platform
-          </p>
-        </div>
+      <div className="h-full overflow-auto">
+        <div className="container mx-auto py-10">Loading videos...</div>
+      </div>
+    );
+  }
 
-        {/* Multi-file Upload Form */}
-        <form onSubmit={handleSubmit(onSubmit)} className="mb-8">
-          <div className="border p-6 bg-card">
-            <h2 className="text-xl font-semibold mb-4">Upload Videos</h2>
-            <div className="space-y-4">
-              {/* Drag and Drop Zone */}
-              <div
-                onDragEnter={handleDragEnter}
-                onDragOver={handleDragOver}
-                onDragLeave={handleDragLeave}
-                onDrop={handleDrop}
-                onClick={() => fileInputRef.current?.click()}
-                className={`
-                  relative border-2 border-dashed p-8
-                  transition-all duration-200 cursor-pointer
-                  flex flex-col items-center justify-center gap-3
-                  ${
-                    isDragging
-                      ? "border-primary bg-primary/5 scale-[1.02]"
-                      : "border-muted-foreground/25 hover:border-primary/50 hover:bg-accent/50"
-                  }
-                `}
-              >
-                <Upload
-                  className={`w-12 h-12 ${
-                    isDragging ? "text-primary" : "text-muted-foreground"
-                  }`}
-                />
-                <div className="text-center">
-                  <p className="text-lg font-medium">
-                    {isDragging
-                      ? "Drop your video files here"
-                      : "Drag & drop video files here"}
-                  </p>
-                  <p className="text-sm text-muted-foreground mt-1">
-                    or click to browse your computer
-                  </p>
-                </div>
-                <Input
-                  {...register("files", {
-                    onChange: handleFileInputChange,
-                  })}
-                  ref={(e) => {
-                    register("files").ref(e);
-                    fileInputRef.current = e;
-                  }}
-                  id="video-files"
-                  type="file"
-                  accept="video/*"
-                  multiple
-                  className="hidden"
-                />
-              </div>
+  return (
+    <div className="container pl-16 mx-auto py-10">
+      <div className="mb-6">
+        <h1 className="text-3xl font-bold">Videos</h1>
+        <p className="text-muted-foreground mt-2">
+          Ready-to-use video clips available in the platform
+        </p>
+      </div>
 
-              {/* Selected Files List */}
-              {selectedFilesList.length > 0 && (
-                <div className="space-y-2">
-                  <Label className="text-sm font-medium">
-                    Selected files ({selectedFilesList.length})
-                  </Label>
+      <MultiFileUploader
+        accept="video/*"
+        uploadEndpoint={ENDPOINTS.UPLOAD_VIDEO}
+        checkFileExists={handleCheckFileExists}
+        onUploadProgress={handleUploadProgress}
+        onUploadComplete={handleUploadComplete}
+        multiple={true}
+        maxConcurrentUploads={3}
+        className="mb-8"
+      />
 
-                  {/* Overall Progress */}
-                  {isUploading && (
-                    <div className="space-y-2 p-4 border bg-muted/50">
-                      <div className="flex justify-between text-sm">
-                        <span className="font-medium">Overall Progress</span>
-                        <span className="text-muted-foreground">
-                          {uploadableCompleted}/{uploadableFilesCount} completed
-                          {uploadableFailed > 0 &&
-                            ` • ${uploadableFailed} failed`}
-                        </span>
-                      </div>
-                      <Progress value={overallProgress} className="h-2" />
-                      <div className="text-xs text-muted-foreground">
-                        {Math.round(overallProgress)}% complete
-                        {uploadStates.length - uploadableFilesCount > 0 &&
-                          ` • ${uploadStates.length - uploadableFilesCount} skipped (already exists)`}
-                      </div>
-                    </div>
-                  )}
-
-                  <div className="border divide-y max-h-60 overflow-y-auto">
-                    {selectedFilesList.map((file, index) => {
-                      const uploadState = uploadStates[index];
-                      const status = uploadState?.status ?? "pending";
-                      const progress = uploadState?.progress ?? 0;
-
-                      return (
-                        <div
-                          key={`${file.name}-${index}`}
-                          className="p-3 hover:bg-accent/50 transition-colors"
-                        >
-                          <div className="flex items-center justify-between mb-2">
-                            <div className="flex-1 min-w-0 flex items-center gap-2">
-                              {status === "uploading" && (
-                                <Loader2 className="h-4 w-4 animate-spin text-primary" />
-                              )}
-                              {status === "completed" && (
-                                <CheckCircle2 className="h-4 w-4 text-green-500" />
-                              )}
-                              {status === "failed" && (
-                                <AlertCircle className="h-4 w-4 text-destructive" />
-                              )}
-                              <div className="flex-1 min-w-0">
-                                <p className="text-sm font-medium truncate">
-                                  {file.name}
-                                </p>
-                                <p className="text-xs text-muted-foreground">
-                                  {(file.size / (1024 * 1024)).toFixed(2)} MB
-                                  {status === "failed" &&
-                                    uploadState?.error &&
-                                    ` • ${uploadState.error}`}
-                                </p>
-                              </div>
-                            </div>
-                            {!isUploading && (
-                              <Button
-                                type="button"
-                                variant="ghost"
-                                size="sm"
-                                onClick={() => removeFile(index)}
-                                className="ml-2 h-8 w-8 p-0"
-                              >
-                                <X className="h-4 w-4" />
-                              </Button>
-                            )}
-                          </div>
-                          {(status === "uploading" ||
-                            status === "completed") && (
-                            <Progress value={progress} className="h-1" />
-                          )}
-                        </div>
-                      );
-                    })}
-                  </div>
-                </div>
-              )}
-
-              <div className="flex gap-2">
-                {fileCount > 0 &&
-                  !isUploading &&
-                  uploadStates.some((s) => s.status === "pending") && (
-                    <Button type="submit" disabled={uploadableFilesCount === 0}>
-                      Upload{" "}
-                      {uploadableFilesCount > 0 &&
-                        `${uploadableFilesCount} file${uploadableFilesCount !== 1 ? "s" : ""}`}
-                    </Button>
-                  )}
-                {fileCount > 0 && isUploading && (
-                  <Button type="button" disabled>
-                    <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                    Uploading...
-                  </Button>
-                )}
-                {fileCount > 0 && !isUploading && (
-                  <Button
-                    type="button"
-                    variant="outline"
-                    onClick={() => {
-                      setSelectedFilesList([]);
-                      setUploadStates([]);
-                      setIsPostUpload(false);
-                      reset();
-                      if (fileInputRef.current) {
-                        fileInputRef.current.value = "";
-                      }
-                    }}
-                  >
-                    Clear all
-                  </Button>
-                )}
-              </div>
-            </div>
-          </div>
-        </form>
-
+      {filteredVideos.length > 0 ? (
         <Table>
           <TableCaption>A list of loaded videos.</TableCaption>
           <TableHeader>
@@ -755,12 +142,11 @@ export const Videos = () => {
             ))}
           </TableBody>
         </Table>
-      </div>
-    );
-  }
-  return (
-    <div className="h-full overflow-auto">
-      <div className="container mx-auto py-10">Loading videos</div>
+      ) : (
+        <div className="text-center py-10 text-muted-foreground">
+          No videos uploaded yet. Upload your first video above.
+        </div>
+      )}
     </div>
   );
 };
